requirements.txt
Django==1.9.5
djangorestframework==3.3.3
gunicorn==19.6.0
requests==2.10.0

